+++ src/simulation/universe/Universe.java
@@ -1,9 +1,6 @@
 package simulation.universe;
 
-import body.SpaceShip;
 import body.interfaces.Body;
-
-import java.util.List;
 
 public interface Universe {
 
@@ -22,7 +19,7 @@ public interface Universe {
      *
      * @param timeStep time step at which to iterate
      */
-    void iteratePhysics(double timeStep);
+    void iteratePhysics(int timeStep);
 
     /**
      *
